# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /api/v1/admin/users/{id}:
    get:
      tags:
      - User Controller
      summary: Получение пользователя по id
      operationId: getUserById
      parameters:
      - name: id
        in: path
        description: Уникальный идентификатор пользователя
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    put:
      tags:
      - User Controller
      summary: Обновление данных пользователя
      operationId: updateUser
      parameters:
      - name: id
        in: path
        description: Уникальный идентификатор обновляемого пользователя
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        description: Данные о пользователе для обновления
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    delete:
      tags:
      - User Controller
      summary: Удаление пользователя
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        description: Уникальный идентификатор удаляемого пользователя
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "204":
          description: No Content
  /api/v1/cards/transfer:
    post:
      tags:
      - User Card Controller
      summary: Перевод между своими картами
      operationId: transferMoney
      requestBody:
        description: Данные для перевода денег
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
...
  /api/v1/cards/{id}/block:
    patch:
      tags:
      - User Card Controller
      summary: Запрос на блокировку своей карты
      operationId: blockRequestFromUser
      parameters:
      - name: id
        in: path
        description: Уникальнеый идентификатор карты для блокировки
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
...
  /api/v1/cards:
    get:
      tags:
      - User Card Controller
      summary: Получение всех своих карт пользователем
      operationId: getAllMyCards
      parameters:
      - name: cardFilter
        in: query
        description: Набор фильтров для карт
        required: true
        schema:
          $ref: "#/components/schemas/CardFilter"
      - name: pageable
        in: query
        description: Настройки Pageable
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardResponse"
  /api/v1/cards/{id}/balance:
    get:
      tags:
      - User Card Controller
      summary: Получение баланса карты
      operationId: getBalance
      parameters:
      - name: id
        in: path
        description: Уникалльный идентификатор карты
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: number
                format: double
  /api/v1/admin/users:
    get:
      tags:
      - User Controller
      summary: Получение всех пользователей
      operationId: getAllUsers
      parameters:
      - name: userFilter
        in: query
        description: Набор фильтров для пользователей
        required: true
        schema:
          $ref: "#/components/schemas/UserFilter"
      - name: pageable
        in: query
        description: Настройки Pageable
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageUserResponse"
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    UserRequest:
      type: object
      properties:
        phone:
          type: string
          description: Номер телефона
          example: "+79606006060"
          pattern: "^\\+?[1-9]\\d{1,14}$"
        email:
          type: string
          description: email пользователя
          example: email@mail.ru
        name:
          type: string
          description: ФИО пользователя
          example: Иванов Иван Иванович
          maxLength: 100
          minLength: 0
        password:
          type: string
          description: Новый пароль пользователя
          example: secret_pass1
          maxLength: 2147483647
          minLength: 8
      required:
      - email
      - name
      - phone
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum:
          - ROLE_USER
          - ROLE_ADMIN
        cardIds:
          type: array
          items:
            type: string
            format: uuid
    TransferRequest:
      type: object
      properties:
        sourceCardId:
          type: string
          format: uuid
          description: id карты списания
          example: 5c87ad40-ca0e-42a7-ae86-c406e776f81e
        targetCardId:
          type: string
          format: uuid
          description: id карты зачисления
          example: dc6070b2-f478-493b-8fb7-91dcedbb7a25
        amount:
          type: number
          format: double
          description: Сумма перевода
          example: 5000
          minimum: 0.01
      required:
      - sourceCardId
      - targetCardId
...
    CardResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cardNumber:
          type: string
        expiryDate:
          type: string
          format: date
        status:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
          - BLOCK_REQUESTED
        balance:
          type: number
        userId:
          type: string
          format: uuid
    CardFilter:
      type: object
      properties:
        status:
          type: string
          description: Статус карты
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
          - BLOCK_REQUESTED
          example: ACTIVE
        userId:
          type: string
          format: uuid
          description: Id пользователя
          example: 329f82e1-6d5a-40d9-91f3-11e5fccd249a
        balance:
          type: number
          format: double
          description: Больше или равно балансу на карте
          example: 1000
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageCardResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/CardResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    UserFilter:
      type: object
      properties:
        role:
          type: string
          description: Роль пользователя
          example: ROLE_ADMIN
    PageUserResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT